+++ breakwater/src/sinks/cli_args.rs
@@ -24,6 +24,10 @@ pub struct SinkCliArgs {
     #[cfg(feature = "vnc")]
     #[clap(flatten)]
     pub vnc_sink: crate::sinks::vnc::VncSinkCliArgs,
+
+    #[cfg(feature = "web")]
+    #[clap(flatten)]
+    pub web_sink: crate::sinks::web::WebSinkCliArgs,
 }
 
 impl SinkCliArgs {
@@ -83,12 +87,6 @@ impl SinkCliArgs {
         // requirements in a `validate` method, which we only invoke when that sink is actually enabled.
         if self.enabled_sinks.contains(&Sink::Ffmpeg) {
             self.ffmpeg_sink
-                .validate()
-                .map_err(|msg| cmd.error(ErrorKind::MissingRequiredArgument, msg))?;
-        }
-        #[cfg(feature = "vnc")]
-        if self.enabled_sinks.contains(&Sink::Vnc) {
-            self.vnc_sink
                 .validate()
                 .map_err(|msg| cmd.error(ErrorKind::MissingRequiredArgument, msg))?;
         }